Replace raw data with Pinia stores
