--- src/components/projects/583/veracruz/data/locations.ts
@@ -1,10 +1,0 @@
-export type Locations = { [k: string]: [number, number] }
-export const locations: Locations = {
-    veracruz: [202, 953],
-    mexicoCity: [119, 947],
-    portAuPrince: [873, 975],
-    washington: [740, 316],
-    colon: [660, 1239],
-    losAngeles: [47, 487],
-    tampico: [153, 859]
-}
