+++ src/components/projects/583/veracruz/types/Location.ts
@@ -1,0 +1,1 @@
+export type Location = [number, number]
