+++ src/main.ts
@@ -1,6 +1,9 @@
-import { createApp } from 'vue'
 import './reset.scss'
 import './main.scss'
+import {createApp} from 'vue'
+import {createPinia} from "pinia";
 import App from './App.vue'
 
-createApp(App).mount('#app')
+const app = createApp(App);
+app.use(createPinia())
+app.mount('#app')
